+++ software/STM32F411_ILI9341_HX711/TouchGFX/gui/src/settingsscreen_screen/SettingsScreenPresenter.cpp
@@ -21,41 +21,41 @@ void SettingsScreenPresenter::Update_Weight_Text(int32_t value)
 	view.Update_Weight_Text(value);
 }
 
-void SettingsScreenPresenter::Update_RawWeight_Text(uint32_t value)
+void SettingsScreenPresenter::Update_RawWeight_Text(uint32_t  value)
 {
 	view.Update_RawWeight_Text(value);
 }
 
-void SettingsScreenPresenter::Update_CalibrationX0_Text(uint32_t value)
+void SettingsScreenPresenter::Update_CalibrationX0_Text(uint32_t  value)
 {
 	view.Update_CalibrationX0_Text(value);
 }
 
-void SettingsScreenPresenter::Update_CalibrationY0_Text(uint32_t value)
+void SettingsScreenPresenter::Update_CalibrationY0_Text(uint32_t  value)
 {
 	view.Update_CalibrationY0_Text(value);
 }
-void SettingsScreenPresenter::Update_CalibrationX1_Text(uint32_t value)
+void SettingsScreenPresenter::Update_CalibrationX1_Text(uint32_t  value)
 {
 	view.Update_CalibrationX1_Text(value);
 }
-void SettingsScreenPresenter::Update_CalibrationY1_Text(uint32_t value)
+void SettingsScreenPresenter::Update_CalibrationY1_Text(uint32_t  value)
 {
 	view.Update_CalibrationY1_Text(value);
 }
-void SettingsScreenPresenter::Cx0_Update(double value)
+void SettingsScreenPresenter::Cx0_Update(uint32_t value)
 {
 	model->Cx0_Update(value);
 }
-void SettingsScreenPresenter::Cx1_Update(double value)
+void SettingsScreenPresenter::Cx1_Update(uint32_t value)
 {
 	model->Cx1_Update(value);
 }
-void SettingsScreenPresenter::Cy0_Update(double value)
+void SettingsScreenPresenter::Cy0_Update(uint32_t value)
 {
 	model->Cy0_Update(value);
 }
-void SettingsScreenPresenter::Cy1_Update(double value)
+void SettingsScreenPresenter::Cy1_Update(uint32_t value)
 {
 	model->Cy1_Update(value);
 }
